third_party/WebKit/LayoutTests/editing/execCommand/switch-list-type-with-orphaned-li.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        [
            '<div contenteditable>^',
                '<ul>',
                    '<li>webkit</li>',
                '</ul>',
                '<li>rocks</li>', // orphaned list child
                '<ul>',
                    '<li>because of</li>',
                '</ul>',
                '<ol>',
                    '<li>you</li>',
                '</ol>',
            '|</div>',
        ].join(''),
        'insertOrderedList',
        [
            '<div contenteditable>',
                '<ol>',
                    '<li>^webkit</li>',
                    '<li>rocks</li>',
                    '<li>because of</li>',
                    '<li>you|</li>',
                '</ol>',
            '</div>',
        ].join('')),
        'Switching multiple lists with an orphaned list child to an ordered list');
</script>
